Added Pattern for Labels + moved some test code.
